+++ GDX/Editor/PropertyDrawers/SerializedDictionaryPropertyDrawer.cs
@@ -1,0 +1,36 @@
+// dotBunny licenses this file to you under the MIT license.
+// See the LICENSE file in the project root for more information.
+
+using GDX.Collections.Generic;
+using UnityEditor;
+using UnityEngine;
+
+namespace GDX.Editor.PropertyDrawers
+{
+    /// <summary>
+    ///     Hides the <see cref="SerializableDictionary{TKey,TValue}"/> for known type arrangements from the inspector window.
+    /// </summary>
+    /// <remarks>
+    ///     In instances where the type is not caught, an empty line will show in the inspector.
+    /// </remarks>
+    [CustomPropertyDrawer(typeof(SerializableDictionary<int, int>))]
+    [CustomPropertyDrawer(typeof(SerializableDictionary<int, string>))]
+    [CustomPropertyDrawer(typeof(SerializableDictionary<int, float>))]
+    [CustomPropertyDrawer(typeof(SerializableDictionary<int, bool>))]
+
+    [CustomPropertyDrawer(typeof(SerializableDictionary<string, int>))]
+    [CustomPropertyDrawer(typeof(SerializableDictionary<string, string>))]
+    [CustomPropertyDrawer(typeof(SerializableDictionary<string, float>))]
+    [CustomPropertyDrawer(typeof(SerializableDictionary<string, bool>))]
+    public class SerializedDictionaryPropertyDrawer : PropertyDrawer
+    {
+        public override float GetPropertyHeight(SerializedProperty property, GUIContent label)
+        {
+            return 0;
+        }
+
+        public override void OnGUI(Rect position, SerializedProperty property, GUIContent label)
+        {
+        }
+    }
+}