+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [1.2.3] - 2021-02-3
 ***"Workplace 2.0"***
 ### Added
+- A Unity serializable dictionary-esque type `GDX.Collections.Generic.SerializableDictionary`.
 - `GDX.Vector3Extensions.DistanceToRay()`
 - `GDX.CapsuleCollider.OutSphereCenters()`
 - `GDX.StringExtensions`